Change UART1 to UART5 and connect 'stdio'.
UART5 on the SAMD21 Curiosity Nano is connected to the USB CDC serial port, which shows up on the host as /dev/ttyACM0 (or something else on other operating systems). This is a better choice to use as the main UART, and SERCOM1 is by convention used as the SPI interface on these dev boards. Also use '_mon_putc()' to send 'stdio' output to the USB CDC port so that we can use functions like 'printf()'.