WIP: add version of code to work with GCC.
GCC and MPLAB X use different names and data structures to access SAMD21 registers. So the new code in 'blinky.c' uses the GCC conventions, and the 'Makefile' invokes GCC to build it. We're borrowing the GCC which is installed as part of the Arduino toolchain, along with all the SAMD21 header files, startup code, and linker scripts. The Arduino IDE must be set up with support for the Arduino M0, a now-retired Arduino board with a SAMD21 chip on it.